Mock keycloak server and redis cache.
For an unauthorised user, the tests were trying to contact the keycloak
server which doesn't exist when the tests are running. I've mocked this
out in the unauthorised tests.
There were also exceptions relating to redis because there was no redis
instance to connect to. These weren't causing anything to fail but they
were annoying. I've mocked out the redis module so these exceptions
don't happen.

I've moved the overrides to the test helper trait and changed the tests
to use this.