Refactor tests to work with the i18n support
The problem was that when the controllers in the tests are created using
their constructor, the calls to the messages api are not replaced with
their actual values. If we inject the controller, it does. So the tests
have had to be changed to expect different text depending on the method
of constructor creation.