Only deploy each version to Docker Hub once
When code is merged to master, build and deploy it to Docker Hub,
using the version tag (e.g 'v123') as the Docker tag.

When a version is deployed to an environment, pull the tagged
Docker image and re-tag it with the environment name.

This means that we can deploy specific version numbers to ECS,
without having to update the task definition in ECS every time.

This process is slightly inefficient because the deployment build
has to pull the image in order to re-tag it, but there doesn't
seem to be an easy way to add a tag in Docker Hub without pulling
and pushing an image.