Added npm i check to the test runner.
The problem was that if this is the first time checking out the project,
node_modules is not populated, sbt test fails because it can't run jest.

This runs npm install first before it runs the tests which makes sure it
has all the dependencies. This will add a slight delay for the tests but
if npm install has already been run, it doesn't take long to run again
so I think we can live with that.