Add tests for the UploadController
These are testing the check to see whether the transfer agreement has
been signed before we allow them to upload. If it has then they can load
the upload page, if not or if they haven't agreed to everything, then
they are redirected to the transfer agreement page.

It's unlikely that we'll get a partially agreed transfer agreement but
it's possible if they hit the API directly so it's worth checking. In
the future we may populate the page with the already filled in values
but seeing as there's no legitimate way to get in this state yet, we can
worry about it later.