Upgrade GraphQL client library
The latest version of the library (0.0.5) serialises optional variables
differently. If they are set to `None`, the library now skips the
variable rather than setting it to `null`.

This makes it possible to remove deprecated variables from the API,
because a variable that has been removed must not be present in the
query, even if it is set to `null`.