Add redirect functionality for once upload is complete
This change redirects the user to the 'processing your records'
placeholder page once the s3 upload has completed. It is done using a
second hidden form on the upload page, the action of which is do render
the redirect page. This hidden form is only submitted once the upload
has completed.

This change still needs tests to be added to it.