Add post build action to Jenkinsfile to run e2e tests.
This change triggers the e2e tests to run on intg after an intg
deployment has been started. There is a potential issue in that the
deployment may not have completely finished and the load balancer
may not be sending traffic to the newly deployed version yet. As a small
workaround to this, the e2e tests file will be modified to include a
'quietPeriod' in which it waits a pre-configured amount of time to build
after a build is triggered.

A more sophisticated solution to this problem may exist, and I will look
into it to see what can be done to fix it.