Add quiet period when triggering e2e test build after deployment
A quiet period means a jenkins job will wait a number of seconds after
being triggered before it begins a build. It is better to use this on
specified builds rather than every time we wish to run the e2e tests. This
may change when a shared jenkins library is created.