Configure Redis cache to cause a failure if Redis is unavailable
By default, the play-redis module is configured to log any errors and
then continue and ignore the error: https://github.com/KarelCemus/play-redis/blob/2.6.1/src/main/resources/reference.conf#L247

This is because the Play cache is just a cache, so it should be safe
(if not performant) for the application to ignore any cache failures.

However, Keycloak uses the Play cache to store OAuth2 state parameters,
so login will not work if the cache is unavailable.

Configuring the cache module to actually fail if it hits an error means
that we get more meaningful error messages. Before, if you tried to log
in while Redis was down, you would get a confusing "State parameter is
different from the one sent in the authentication request" error when
you submitted the login form. Now, you can't even get to the login page
if Redis is down, and the stack trace is clearer because one of the
causes is `redis.actors.NoConnectionException$: No Connection established`.