Configure local S3 upload endpoint
Configure the upload step to send files to a local S3 emulator instead
of to the real S3 endpoint. This makes it possible to do entirely local
development without relying on AWS for S3 and all the backend checks.

This lets us run an entirely local development environment, with backend
checks saving the results to the local database, which should make it
easier to develop the failure path pages without waiting for the real
checks to finish.

The fake S3 upload requires a Cognito identity pool ID, so use a local
emulated version to fetch this ID.

This commit sends _all_ requests to the local S3 emulator. A later commit
will make the endpoint configurable, so that we can easily switch between
the real S3 and the fake S3 in the development environment.