Add set up for record progress update on records page
This is a WIP and mostly set up work for getting the progress of
AVFileMetadata on records processing page upon page load. The rest of
the work on this will have to wait until the generated GraphQL pr has
been merged.

This is mostly just draft work so front-end may not completely match
wireframes etc.