Test the folder selection method through the upload method
This makes the test more thorough, because it also tests the
code which finds the element to add the event listener to.

Ideally, we would extract the target from the event that was
called rather than from the element itself. This turned out to
be very hard to replicate in the tests, because the Event
cannot be constructed directly with the right target, and the
Event interface has many methods which would need to be mocked.

Co-authored-by: Nigel Lutale <nigel.lutale@nationalarchives.gov.uk>