Don't throw generic error
In the error handling handleError function, we are now checking the type
of the error, in this case, if it's a login error and doing different
things depending on this.
This try catch is turning any custom errors into generic errors which
isn't really needed so I'll take it out.