Remove node specification for Jenkins post-build steps
This should fix some Jenkins deadlocks that happen when too many builds
are waiting for a Jenkins primary node to become available.

Specifying the "master" node is unnecessary because the pipeline already
uses the "master" node by default. Specifying it again in the post-build
step means the build waits for a second primary node to become available,
which can cause a deadlock if multiple builds are in the same state.