Update drag and drop so that it detects multiple folders
With the previous implementation, the drag and drop dropped files onto an enlarged input button, which meant that it did not detect multiple folders and depending on browser, files.

This new implementation should detect any item dropped on it. This means that we can show an error message to the user if they unitentionally drop multiple files or folders.

This is quite a large change because the drop event structures files in a nested tree, whereas the imput button structed them as an array of paths.

Co-authored-by: Nigel Lutale <nigel.lutale@nationalarchives.gov.uk>
Co-authored-by: Sam Palmer <sam.palmer@nationalarchives.gov.uk>