Add 'accept' parameter to file input to fix Safari bug
Specify an arbitrary type in the 'accept' attribute to work around a bug
in Safari 14.0.1, which does not let the user browse for files if the
'accept' attribute is missing. The actual value of the attribute is
ignored because 'webkitdirectory' is specified. It just needs to be present
to fix the Safari bug.

Other bug reports:

- https://stackoverflow.com/q/64843459/679924
- https://discussions.apple.com/thread/252051818