Extend the connection timeout.
It looks like the connection timeout is causing the problems with large
file uploads. At least, when I make it very short, I can reliably make
the problem happen and Tom is getting errors after 2 minutes which is
the default timeout. I've upped this to 20 minutes to see if this will
help. It's a bit trial and error at the moment so we'll see.