Create transfer complete page.
We already had a placeholder but this is implementing the design for the
page.
We're not displaying a specific error if triggering the export fails so
I've removed the parameter and the test for it.
I've also renamed everything from transfer confirmation to transfer
complete to match with the title.