Hide the upload form rather than the whole container.
If an upload had already occurred for the consignment, it was hitting
the handleUploadError function. This was hiding the entire container for
the upload form including the error message so we were just seeing a
blank page.
I've changed it to hide the <form> element and show the error and this
is now working as it used to.