Change date format. Remove file appender
I've removed the file appender because we're not using it.

The date format has been changed because we're trying to sort out
multiline logging using https://docs.docker.com/config/containers/logging/awslogs/#awslogs-datetime-format
The problem is that the millisecond option %L matches to .123 with the
mandatory leading dot and there's no way to get the format to accept
,123
There's a github issue here https://github.com/moby/moby/issues/36796
which doesn't have any recent activity so the only way to use this is to
change the format to use a dot instead of the default comma.