Updated heading levels and spacing
Changed some paragraphs with class of a heading into headings and updated all headings to be the same level (making sure they have the correct hierarchy) matching the GDS guidelines and templates.
Also added the missing divs on the dashboard page and changed 'Transfer Complete' to 'Transfer complete' to be consistent with our wording.