Convert Keycloak SSO silent auth login page to Play-rendered page
This was a static HTML page with some inline JavaScript. Move it to a
regular Play page, which loads a separate JavaScript. The functionality
is unchanged.

We would like to add a Content Security Policy which bans inline JavaScript,
since it is a cross-site scripting vector. The CSP would have blocked
this legitimate Keycloak inline JS, so it needs to be moved to a separate
script.