Move govuk-frontend initialization into main TypeScript module
Removed the inline script which used to initialize govuk-frontend.

The two methods of initialization are equivalent:
https://frontend.design-system.service.gov.uk/importing-css-assets-and-javascript/#javascript

The advantage of doing the initialization in the TypeScript module instead
is that it removes inline JavaScript, which means we can prevent all
inline JavaScript using a Content Security Policy. This reduces the
risk of cross-site scripting attacks. The CSP will be added in a
later commit.