Add webkitRelativePath to mock File objects to better mimic actual File objects
According to MDN, File objects should have a webkitRelativePath property.
It's been fine not to have one on our mock file objects so far, as we haven't been testing the submit event listener's function
(which calls this property to get the parent folder) but when we do, we'll need it.
https://developer.mozilla.org/en-US/docs/Web/API/File