Show 500 page if the parent folder is missing.
The check for the parent folder is a check to make sure that the upload
has been completed before trying to access the records results. The
problem is this isn't being checked if allChecksSucceeded returns false.
This returns false if the file checks haven't happened yet so what you
end up with is a page saying your file checks have failed even before
you've uploaded any files.
I've moved the get or throw line up a bit so we get the 500 page now as
we used to.