Set tabindex to 0
The other error divs are shown with javascript and because of this, the
browser will auto-focus on them which means the screen reader
automatically read them out.
We can show this error with javascript so the tabindex needs to be set
to zero so the user can tab to the error and the screen reader will read
it out.