Switch code so that it checks folder has files before saving array
After the checks for a non-folder and multiple items are done, there is still the possibility that the folder could be empty. Before this commit, it was: first, assigning the potentially empty array (folder) to the form, checking whether the array (folder) was full or empty and then assigning an empty array to the form if the user indeed dropped an empty folder. This would mean that if the folder was empty, it would be assigning the empty folder twice (which is redundant). This commit checks if the folder is empty before any assignment takes place.