Relocated messages into views
The messages file was becoming long and difficult to work with so we are moving all of the text directly into the views. This might make it harder to add translation to TDR but we don't believe we will need to do that in the near future. This change may be especially helpful when we add text-heavy pages like the help guide.

We will add an architectural decision record about this change.

Also updated the wording from the newest review as noted in TDR-1135.