Add preferred algorithm.
This is to prevent the error
`Preferred JWS algorithm: null not available`
which we're getting from the frontend. From debugging the pac4j code and
looking at the keycloak documentation, the algorithm Keycloak is using
is RS256. I've tested this locally and all of the auth is still working
as expected.