Enabled redirect filter
Enabling this enables strict transport security.
https://www.playframework.com/documentation/2.8.x/RedirectHttpsFilter

This only runs in production mode so won't affect local development and
so can go into the base config file.

This means that all http requests get a 308 response. This includes the
load balancer health checks so we need to change the expected response
code for the health check to 308 but this will be done in the terraform
repo.