Move duplicated progress bar code into one function
This was initially recommended by Tom, and implemented by Rooha, and then removed due to the fact that using arrow functions,
binds 'this' to the object that calls it. I added it back because after looking through the code, the change makes sense to me;
we DO want this to be bound to the object that calls it: 'ClientFileProcessing'. Here is what Mozilla have to say about not using
arrow functions; I don't think this code fits these examples
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/Arrow_functions