Updated format and password question
The question about resetting the password did not make much sense for private beta. To avoid confusion, we just tell the user to contact us here. The user will either click the link on the sign in page or come to the FAQ page. No need to redirect the user to the sign in page.

Fixed typo CSV and updated formatting.