Change the file checks continue button from link to a button.
The button needs to be disabled when the file checks are in progress and
you can't disable a link without doing some hacky things like removing
the href so I've converted this to a form with a button and added the
disabled attribute. I've updated the js to remove the disabled attribute
and changed the tests to check.