Add call to API endpoint to change the upload status to 'Completed'
Use dynamic imports for the generated graphql document imports.

After upgrading the generated graphql project to use webpack 5, we
started getting errors because the imports coming from the
tdr-generated-graphql project were undefined.

Doing this makes them work again. I don't know why. There is probably
something we can do in the generated graphql project to make it a bit
more sane but for now, this will unblock people working on the front end
and we can revisit it later.

Co-authored-by: Sam Palmer <sam.palmer@nationalarchives.gov.uk>