Revert dynamic imports for the generated graphql library.
I've downgraded webpack in the generated-grapqhl library so we don't
need the dynamic imports any more.
We're now importing them as we used to. The mjs changes in the webpack
config are needed to fix an error from the graphql-tag library. It's a
small enough change that I don't think we need to downgrade the
graphql-tag library.