Fix wrong assertion in file check test
.getAttribute() returns the value of the attribute or null if the attribute doesn't exist.
since the value of 'hidden' is nothing, it returns the value ''. Since empty strings are falsy and null is falsy
the previous assertion '.toBeFalsy()' would always pass. This test expects the hidden attribute to NOT exist
and therefore we would expect .getAttribute() to return null so '.toBeNull()' was added instead.