Change the file and metadata process
When the upload starts, StartUpload is called which updates the parent
folder name and sets the consignment status to upload in progress.

Then the metadata is extracted and sent to the API in batches. Entries
in the file table are created at this point and the file IDs are sent
back with a sequence id sent in the original request. These are then
matched up and the upload continues as normal.