Replace map with async function with for loop.
We were mapping over the batches with an async function inside map but
this is causing it to make requests to the API almost simultaneously and
what we want is for them to be sent in sequence.
I've replaced the `map` call with an old fashioned for loop and this is
now sending the requests sequentially.