Make only the % get read out & add aria polite to not interrupt instructions
Each time the page was updated, it was updated with 'Uploading records n %', which meant that the screen reader was reading
the whole thing. This commit adds span element around 'n%' so that only that gets read out each update. aria polite was addded
so that it does not interrupt the 'Do not close your browser...etc.' instruction that's read out first when page is revealed.