Don't set the requested url in restore method.
I tried this by going directly to the series page while logged out
What seems to be happening is:

1. You go to the series page. You're redirected to the login so `save()`
   is called and `series` is set in the session.
2. You login and `restore()` is called. `series` is retrieved from the
   session and then the `REQUESTED_URL` value is set to empty string.
   You're redirected to the series page.
3. You click back. The previous page was the keycloak url so you don't
   hit the front end app at all so there's no chance to call the `save`
   method. `REQUESTED_URL` is still set to empty string.
4. You hit the Keycloak url, you're still logged in so Keycloak
   redirects you back to the front end app via /callback. The `restore()`
   method is called. The value of `REQUESTED_URL` is still empty string.
   This is retrieved and we then try to cast it to `FoundAction`. You
   can't cast `String` to `FoundAction` so you get an error.

I've changed this so we don't set the value of `REQUESTED_URL` in the
`restore()` method and everything seems to still be working as it was.

If you go directly to the transfer-agreement page while logged out, it
will redirect you to the login page. After login, you'll be redirected
to the transfer agreement page. If you then click back, it will redirect
you to Keycloak, find you're still logged in and redirect you back to
the transfer agreement page. From the users point of view, this means
that the back button keeps them on the same page but considering that's
the page they went to in the first place, I think this is fine.

If you're already in the app and you go from '/' -> 'series' -> `transfer-agreement`
and your session times out, then you click back. You'll be redirected to
the login page and then after login, be redirected to the `series` page
which seems like sensible behaviour.