Add canned ACL.
The owner of the objects will be the Cloudfront OAI which means by
default, anything in the account that isn't Cloudfront has no access to
it which is not useful for the download files or antivirus lambda.

I've set this so that the bucket owner can read from this bucket and
I've set the bucket policy for that bucket to only allow the roles from
the download files and antivirus lambdas.