Move the get cookies url into the upload handler.
This will prevent a user being able to select and upload files before
the cookies endpoint has returned the signed cookies.

What does happen now is that there is a noticeable delay after you click
upload when the lambda has to start from cold but we should be able to
get around this if it causes us problems in the future.