Add section about icognito mode and local development.
Because the upload.tdr-integration url uses cookie authentication and
icognito mode blocks third party cookies, it doesn't work by default
when running the front end locally.

This explains that and adds some links to the browser documentation to
add an exception. I've tested this locally with an added exception and
it works.