Make the status 200 when navigating back to Transfer Agreement Page
There was an issue where if you tried to submit an incomplete form, submitted a complete form (after the
page had reloaded with the error messages), clicked 'Continue' to go to the 'Upload' page and clicked the back
button of your browser, it would show the user the TA page with the error messages, instead of the 'already confirmed'
page. Making the status code, for POST requests that occur when the Transfer Agreement status has been set to 'Completed', 200,
remedies this and loads the 'already confirmed' page.