Remove options from form to mimic incomplete submission, instead of making them 'false'
I was initially under the wrong impression that a 'false' value in the form meant that the option was 'not selected'; it doesn't.
'false' simply means 'the opposite value', it used to make sense when we had 'Yes' & 'No' options as, they would map to 'true' & 'false' respectively.
So actually not providing the option in the form, mimics not selecting the option on the TA page.