Pass 'checked' into inputSingleCheckbox instead of 'true'
Previously, the value for 'Symbol('_checkedOption') ->' (which is passed to the inputSingleCheckBox Partial) accepts was changed
from 'true' to 'checked' as a way of simplifying the codebase. Since the final transfer confirmation view relies on this partial
and the partial was not changed, it caused a silent bug where if you submitted an partially complete form, the option that you selected was not
checked by default i.e. you have to select it again as well as the option you didn't select.