Remove form from test
This test is mistakenly passed a form with false values in order to test an incomplete form submission; it shouldn't be passed a form
at all. This commit removes the form. This commit also implements the check for the form options' attributes mentioned in a earlier commit
on this branch