Genericise 'rejectUserItemSelection' & warning messages
Since the number of warning messages is going to increase, it makes sense to remove them from rejectUserItemSelection (where they were fixed) and
pass in the one that shouldn't be hidden (should be displayed) so that everything else can be hidden. The error message now accompanies it for better debugging.
It also made sense to make the warning message names more generic so that they make sense for files and folders