Use base alpine image with jdk 15 installed
Tried upgrading busybox on the openjdk image but it seems to be
on an older version of alpine and there are still vulnerabilities in the
latest version of busybox for that version.

Instead of using the openjdk image, use the
standard alpine image and installed openjdk 15 from the alpine edge
repository.