Get the consignment type from API
Rather than getting this from the token, we're now getting it from the
API. This will help if we accidentally have a user who is a user of both
groups.

There are two exceptions. The series page which is only accessible by a
standard user but we don't have a consignment at that point. The other
one is the submit endpoint to create the consignment on the homepage
which is only accessibly by judgment users. The token needs to be
checked for this but there's not much we can do about that.