Add tests for the new functionality.
This checks that it rejects an empty directory if there are no files in
the upload.
It checks to make sure that empty directories don't show up in the
number of files uploaded.

There is a test for the metadata processing to check they are filtered
out of the file inputs correctly.