Stop throwing errors.
It looks like that some of our error handling is causing the javascript
to throw an error when there is a temporary failure of the upload which
is corrected by the SDK so we shouldn't be throwing an error.

We're using thrown errors for normal validation which is not what
they're for really, they should be for exceptional failures.
 The reason we did this originally is that we would end up with a load of
branching if/else for the errors if we didn't throw Errors for validation.

Now that throwing errors is causing problems, we're going to have to
live with the branching if/else statements until Typescript sorts out a
decent function style of error handling like Scala's Either.