Remove DTS bundle plugin.
Dependabot is trying to upgrade minimist to the latest stable version as
there's an alert on it but it can't because dts-bundle is using an old
version.

The project itself seems dead. It hasn't had a commit since 2017. I'm
not sure why we added it but we don't seem to need it. We still get the
typescript mapping files in the browser which is all we need.
We aren't publishing this anywhere so we don't need d.ts files creating.