Replace instances of 'directory' with 'folder'
All the places that were named “Directory“ to “Folder“ where it made sense were renamed
but there are some remaining cases where I wasn’t sure if we should. For example, we create objects with names like
'emptyDirectoryEntry' and nestedDirectoryEntry in order to closely mimic the DOM API’s DirectoryEntry object,
as well as 'isDirectory' but this is an inherited property from an interface based on a DOM object
so it would have been confusing if we renamed it.