Change the frontend to allow a /27 subnet mask.
The private subnet which the apps sit in is on a /28 which is only 11
usable addresses. We're hitting this limit at the moment because there's
two database instances running but even when we drop the original
databases, we still get problems if there are two deployments and the
bastion is running so we need change this.

Play needs every possible address from the load balancer health checks
so this is every address from two AZs worth of /27 IPs.