Make it so that you have to explicitly pass in form to FormTester
When you submit some forms (ones with radios or checkboxes usually) without selecting values,
what you get on the other end is an empty 'Map()', therefore the default 'optionsSelected' parameter
was set to an empty Map, to indicate 'no values', but the closure metadata form for instance has input boxes
Which always appear in the Map (even if the value is left empty because it seems to turn an empty value into an
empty string), it also has values which are set by default and therefore you can not pass an empty Map.
To avoid accidental omissions/mistakes, the default empty Map has been removed.