Add 'files' only to 'selected cache'
Not adding 'folders' to the selected cache because simplifies the code.

Also it means that in the future it allows for the adding of additional metadata against folders, should this be needed. Current assumption is that additional metadata is not stored against folders.

An individual folder state is calculated 'on the fly' when nodes are rendered on the page. This may cause some performance issues should the be a large number of folders to display.