Version bump components
It turns out that you can't delete a package and republish over the same
version after a certain amount of time. I was trying to clear out the
older versions of the tdr-components library and got rid of everything.
Long story short, this is the next available version which I've
published manually to get the build working.