Update download metadata report.
Originally, this had the filename hard coded in at the beginning of the
report using the original file path. The API has been updated now to
store the file name and the file type in the metadata table and the
relevant file properties now have an ordering on them so this report can
be entirely data driven. I've updated the tests as well.