Skip closure type confirmation if already confirmed.
This does a few things.

If all files are already confirmed as closed, the closure status page is
skipped.

If you go to the closure status page anyway, the checkbox is checked and
disabled and the submit button is changed to a link which takes you to
the custom metadata page.

I've put the logic for whether all files are closed or not in a separate
method so we can change the logic easily.