Prevent storing of cache on Metadata form
If you partially submit the metadata form first, then fill it in completely, submit it (so that you go to the review page ) and then go back
you were still seeing the error messages even though the form was filled in. The uncache() method was added so that it does not cache the page.
The solution works on Firefox but on Chrome (probably a bug), when you click back you get an 'ERR_CACHE_MISS' message telling you that you need
to reload the page. If you ignore this message (choose not to reload the page) and instead go back, it will show you a form without your changes
(even though they were made in the DB); you need to reload the page in order to see the changes.