Remove the call to update consignment status.
We were sending a call to the API to update the consignment status
without sending a status.

The API would then calculate the status based of the existing file
Upload statuses and set the consignment status accordingly.

This is now being done in the backend checks so it can be removed from
here.