Remove duplicate ids and aria-hidden.
The automated accessibility testing tools were flagging duplicate ids
and the aria-hidden attribute on the file upload.

I've changed one of the ids to progress-error-summary-title. Luckily,
this hasn't broken any of the javascript because it's all searching for
the parent div. I've updated the id in the test html as well and they're
still passing so we should be good.

The upload page still looks the same and works the same without the
aria-hidden attribute so I think we can leave that off.